+++ components/IMU/src/Mahony_Filter.cpp
@@ -11,7 +11,7 @@
 namespace flyhero
 {
 
-Mahony_Filter::Mahony_Filter(double kp, double ki)
+Mahony_Filter::Mahony_Filter(float kp, float ki)
         : MAHONY_KP(kp), MAHONY_KI(ki)
 {
     this->last_time = 0;
@@ -27,7 +27,7 @@ Mahony_Filter::Mahony_Filter(double kp, double ki)
 // expects gyro data in deg/s, accel data in multiples of g
 void Mahony_Filter::Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::Euler_Angles &euler)
 {
-    double delta_t;
+    float delta_t;
 
     if (this->last_time == 0)
     {
@@ -37,12 +37,12 @@ void Mahony_Filter::Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::
     } else
     {
         int64_t tmp = esp_timer_get_time();
-        delta_t = (tmp - this->last_time) * 0.000001;
+        delta_t = (tmp - this->last_time) * 0.000001f;
 
         this->last_time = tmp;
     }
 
-    double recip_norm;
+    float recip_norm;
 
     IMU::Sensor_Data gyro_rad;
     gyro_rad.x = gyro.x * Math::DEG_TO_RAD;
@@ -52,7 +52,7 @@ void Mahony_Filter::Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::
     IMU::Sensor_Data v, error;
 
     // normalise accelerometer measurement
-    recip_norm = 1 / std::sqrt(accel.x * accel.x + accel.y * accel.y + accel.z * accel.z);
+    recip_norm = 1 / sqrtf(accel.x * accel.x + accel.y * accel.y + accel.z * accel.z);
     accel.x *= recip_norm;
     accel.y *= recip_norm;
     accel.z *= recip_norm;
@@ -88,24 +88,24 @@ void Mahony_Filter::Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::
     gyro_rad.y += this->MAHONY_KP * error.y + this->MAHONY_KI * error_integral.y;
     gyro_rad.z += this->MAHONY_KP * error.z + this->MAHONY_KI * error_integral.z;
 
-    double qa = this->quaternion.q0;
-    double qb = this->quaternion.q1;
-    double qc = this->quaternion.q2;
+    float qa = this->quaternion.q0;
+    float qb = this->quaternion.q1;
+    float qc = this->quaternion.q2;
 
-    this->quaternion.q0 += 0.5 * delta_t * (-qb * gyro_rad.x - qc * gyro_rad.y
-                                            - this->quaternion.q3 * gyro_rad.z);
-    this->quaternion.q1 += 0.5 * delta_t * (qa * gyro_rad.x + qc * gyro_rad.z
-                                            - this->quaternion.q3 * gyro_rad.y);
-    this->quaternion.q2 += 0.5 * delta_t * (qa * gyro_rad.y - qb * gyro_rad.z
-                                            + this->quaternion.q3 * gyro_rad.x);
-    this->quaternion.q3 += 0.5 * delta_t * (qa * gyro_rad.z + qb * gyro_rad.y
-                                            - qc * gyro_rad.x);
+    this->quaternion.q0 += 0.5f * delta_t * (-qb * gyro_rad.x - qc * gyro_rad.y
+                                             - this->quaternion.q3 * gyro_rad.z);
+    this->quaternion.q1 += 0.5f * delta_t * (qa * gyro_rad.x + qc * gyro_rad.z
+                                             - this->quaternion.q3 * gyro_rad.y);
+    this->quaternion.q2 += 0.5f * delta_t * (qa * gyro_rad.y - qb * gyro_rad.z
+                                             + this->quaternion.q3 * gyro_rad.x);
+    this->quaternion.q3 += 0.5f * delta_t * (qa * gyro_rad.z + qb * gyro_rad.y
+                                             - qc * gyro_rad.x);
 
     // normalise quaternion
-    recip_norm = 1 / std::sqrt(this->quaternion.q0 * this->quaternion.q0 +
-                               this->quaternion.q1 * this->quaternion.q1 +
-                               this->quaternion.q2 * this->quaternion.q2 +
-                               this->quaternion.q3 * this->quaternion.q3);
+    recip_norm = 1 / sqrtf(this->quaternion.q0 * this->quaternion.q0 +
+                           this->quaternion.q1 * this->quaternion.q1 +
+                           this->quaternion.q2 * this->quaternion.q2 +
+                           this->quaternion.q3 * this->quaternion.q3);
 
     this->quaternion.q0 *= recip_norm;
     this->quaternion.q1 *= recip_norm;
@@ -115,21 +115,21 @@ void Mahony_Filter::Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::
     // convert quaternion to euler
     // https://en.wikipedia.org/wiki/Conversion_between_quaternions_and_Euler_angles#Quaternion_to_Euler_Angles_Conversion
 
-    double q2_sqr = this->quaternion.q2 * this->quaternion.q2;
+    float q2_sqr = this->quaternion.q2 * this->quaternion.q2;
 
-    double t0 = +2.0 * (this->quaternion.q0 * this->quaternion.q1 + this->quaternion.q2 * this->quaternion.q3);
-    double t1 = +1.0 - 2.0 * (this->quaternion.q1 * this->quaternion.q1 + q2_sqr);
-    euler.roll = std::atan2(t0, t1) * Math::RAD_TO_DEG;
+    float t0 = +2.0f * (this->quaternion.q0 * this->quaternion.q1 + this->quaternion.q2 * this->quaternion.q3);
+    float t1 = +1.0f - 2.0f * (this->quaternion.q1 * this->quaternion.q1 + q2_sqr);
+    euler.roll = atan2f(t0, t1) * Math::RAD_TO_DEG;
 
-    double t2 = +2.0 * (this->quaternion.q0 * this->quaternion.q2 - this->quaternion.q3 * this->quaternion.q1);
-    t2 = ((t2 > 1.0) ? 1.0 : t2);
-    t2 = ((t2 < -1.0) ? -1.0 : t2);
-    euler.pitch = std::asin(t2);
+    float t2 = +2.0f * (this->quaternion.q0 * this->quaternion.q2 - this->quaternion.q3 * this->quaternion.q1);
+    t2 = ((t2 > 1.0f) ? 1.0f : t2);
+    t2 = ((t2 < -1.0f) ? -1.0f : t2);
+    euler.pitch = asinf(t2);
     euler.pitch *= Math::RAD_TO_DEG;
 
-    double t3 = +2.0 * (this->quaternion.q0 * this->quaternion.q3 + this->quaternion.q1 * this->quaternion.q2);
-    double t4 = +1.0 - 2.0 * (q2_sqr + this->quaternion.q3 * this->quaternion.q3);
-    euler.yaw = std::atan2(t3, t4) * Math::RAD_TO_DEG;
+    float t3 = +2.0f * (this->quaternion.q0 * this->quaternion.q3 + this->quaternion.q1 * this->quaternion.q2);
+    float t4 = +1.0f - 2.0f * (q2_sqr + this->quaternion.q3 * this->quaternion.q3);
+    euler.yaw = atan2f(t3, t4) * Math::RAD_TO_DEG;
 }
 
 void Mahony_Filter::Reset()
